[feat] 카카오 로그인 구현 완료

+++ app/src/main/java/com/example/perfumeproject/di/AuthManager.kt
@@ -33,7 +33,7 @@ class AuthManager @Inject constructor(val sharedPrefs: SharedPrefs) {
             sharedPrefs[EXPIRE_KEY] = value
         }
 
-    var id: Int
+    var id: Long
         get() {
             return sharedPrefs[ID, 0] ?: 0
         }
